Fixed configuration to use a default value if parameter is not added to request

+++ src/main/java/com/calculator/representations/ResultRepresentation.java
@@ -4,16 +4,16 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ResultRepresentation { //POJO: plain old java object
 
-    private float result;
+    private Double result;
     private String operation;
-    private float value1;
-    private float value2;
+    private Double value1;
+    private Double value2;
 
     public ResultRepresentation() {
         //Jackson deserialization
     }
 
-    public ResultRepresentation(float result, String operation, float value1, float value2) {
+    public ResultRepresentation(Double result, String operation, Double value1, Double value2) {
         this.result = result;
         this.operation = operation;
         this.value1 = value1;
@@ -21,7 +21,7 @@ public class ResultRepresentation { //POJO: plain old java object
     }
 
     @JsonProperty
-    public float getResult() {
+    public Double getResult() {
         return result;
     }
 
@@ -31,12 +31,12 @@ public class ResultRepresentation { //POJO: plain old java object
     }
 
     @JsonProperty
-    public float getValue1() {
+    public Double getValue1() {
         return value1;
     }
 
     @JsonProperty
-    public float getValue2() {
+    public Double getValue2() {
         return value2;
     }
 }
